backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
asyncpg==0.29.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
aiosqlite==0.20.0
httpx==0.27.0
